requirements.txt
# CORE
fastapi>=0.118.0,<0.119.0
asyncpg>=0.30.0,<0.31.0
sqlalchemy>=2.0.43,<2.1.0
# SECONDARY
python-jose[cryptography]>=3.5.0,<4.0.0
pydantic>=2.11.9,<2.12.0
python-decouple>=3.8,<4.0
uvicorn>=0.37.0,<0.38.0
alembic>=1.16.5,<1.17.0
# DEV
redis>=6.4.0,<7.0.0
celery>=5.5.3,<6.0.0
itsdangerous>=2.2.0,<3.0.0
# pytest>=7.4.0,<8.0.0
# pytest-asyncio>=0.22.0,<0.23.0
# httpx>=0.24.0,<0.25.0